move cloud_function to functions/aou
