+++ .github/workflows/release_latest_version.yml
@@ -27,7 +27,7 @@ jobs:
           # Maps tcp port 5432 on service container to the host
           - 5432:5432
     env:
-      MODULES: api ui cloud_function
+      MODULES: api ui functions/aou
 
     steps:
       - name: prepare sources
