--- .vscode/launch.json
@@ -1,14 +1,0 @@
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "Python: Deployment File",
-            "type": "python",
-            "request": "launch",
-            "program": "ops/cli.py",
-            "cwd": "${workspaceFolder}",
-            "args": ["deploy"],
-            "console": "integratedTerminal"
-        }
-    ]
-}
