+++ functions/aou/module.mk
@@ -1,4 +1,4 @@
-# Makefile for the cloud_function module
+# module.mk for the functions/aou module
 
 REQUIRED_PYTHON_ENVIRONMENT := dev-requirements.txt
 include $(MAKE_INCLUDE_DIR)/modules.mk
